@@ -1,26 +1,24 @@
 package gueron
 
 import (
-	"context"
+	"database/sql"
 	"errors"
+	"log/slog"
 	"testing"
 	"time"
 
 	"github.com/benbjohnson/clock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/vgarvardt/gue/v5"
-	"github.com/vgarvardt/gue/v5/adapter"
-	dbTest "github.com/vgarvardt/gue/v5/adapter/testing"
-	"github.com/vgarvardt/gue/v5/adapter/zap"
+	"github.com/vgarvardt/gue/v6"
+	"go.uber.org/zap/exp/zapslog"
 	"go.uber.org/zap/zaptest"
 )
 
 func TestNewScheduler(t *testing.T) {
-	connPool := new(dbTest.ConnPool)
-	logger := zap.New(zaptest.NewLogger(t))
+	logger := slog.New(zapslog.NewHandler(zaptest.NewLogger(t).Core()))
 
-	s, err := NewScheduler(connPool, WithHorizon(10*time.Minute), WithLogger(logger))
+	s, err := NewScheduler(nil, WithHorizon(10*time.Minute), WithLogger(logger))
 	require.NoError(t, err)
 	s.
 		MustAdd("@every 1m", "foo", nil).
@@ -46,9 +44,7 @@ func TestNewScheduler(t *testing.T) {
 }
 
 func Test_schedulesHash(t *testing.T) {
-	connPool := new(dbTest.ConnPool)
-
-	s1, err := NewScheduler(connPool)
+	s1, err := NewScheduler(nil)
 	require.NoError(t, err)
 	s1.MustAdd("@every 1m", "foo", nil).
 		MustAdd("*/3 * * * *", "bar", nil).
@@ -57,7 +53,7 @@ func Test_schedulesHash(t *testing.T) {
 	hash1 := s1.schedulesHash()
 	require.NotEmpty(t, hash1)
 
-	s2, err := NewScheduler(connPool)
+	s2, err := NewScheduler(nil)
 	require.NoError(t, err)
 	s2.MustAdd("*/3 * * * *", "bar", nil).
 		MustAdd("@hourly", "bar", []byte(`{"foo":bar}`)).
@@ -67,7 +63,7 @@ func Test_schedulesHash(t *testing.T) {
 	require.NotEmpty(t, hash2)
 	assert.Equal(t, hash1, hash2)
 
-	s3, err := NewScheduler(connPool)
+	s3, err := NewScheduler(nil)
 	require.NoError(t, err)
 	s3.MustAdd("*/3 * * * *", "bar", nil).
 		MustAdd("@hourly", "bar", []byte(`{"foo":bar}`)).
@@ -78,7 +74,7 @@ func Test_schedulesHash(t *testing.T) {
 	assert.NotEqual(t, hash1, hash3)
 
 	randomQueueName := time.Now().Format(time.RFC3339Nano)
-	s4, err := NewScheduler(connPool, WithQueueName(randomQueueName))
+	s4, err := NewScheduler(nil, WithQueueName(randomQueueName))
 	require.NoError(t, err)
 	s4.MustAdd("*/3 * * * *", "bar", nil).
 		MustAdd("@hourly", "bar", []byte(`{"foo":bar}`)).
@@ -94,9 +90,9 @@ func Test_cleanupScheduledLeftovers(t *testing.T) {
 		t.Skip("skipping test in short mode")
 	}
 
-	ctx := context.Background()
-	pool := dbTest.OpenTestPoolPGXv5(t)
-	logger := zap.New(zaptest.NewLogger(t))
+	ctx := t.Context()
+	pool := openTestPoolPGXv5(t)
+	logger := slog.New(zapslog.NewHandler(zaptest.NewLogger(t).Core()))
 	now := time.Now()
 	queue := now.Format(time.RFC3339Nano)
 
@@ -109,13 +105,13 @@ func Test_cleanupScheduledLeftovers(t *testing.T) {
 	s.MustAdd("@every 12h", "dummy-job", nil)
 
 	// no jobs yet, nothing to clean up
-	tx1, err := pool.Begin(ctx)
+	tx1, err := pool.BeginTx(ctx, nil)
 	require.NoError(t, err)
 
 	err = s.cleanupScheduledLeftovers(ctx, tx1)
 	require.NoError(t, err)
 
-	err = tx1.Commit(ctx)
+	err = tx1.Commit()
 	assert.NoError(t, err)
 
 	// schedule several jobs
@@ -146,13 +142,13 @@ func Test_cleanupScheduledLeftovers(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, j2.ID, jLocked.ID)
 
-	tx2, err := pool.Begin(ctx)
+	tx2, err := pool.BeginTx(ctx, nil)
 	require.NoError(t, err)
 
 	err = s.cleanupScheduledLeftovers(ctx, tx2)
 	require.NoError(t, err)
 
-	err = tx2.Commit(ctx)
+	err = tx2.Commit()
 	require.NoError(t, err)
 
 	err = jLocked.Error(ctx, gue.ErrDiscardJob("test"))
@@ -160,7 +156,7 @@ func Test_cleanupScheduledLeftovers(t *testing.T) {
 
 	jLockedAgain, err := s.gueClient.LockJobByID(ctx, j2.ID)
 	require.Error(t, err)
-	assert.True(t, errors.Is(err, adapter.ErrNoRows))
+	assert.True(t, errors.Is(err, sql.ErrNoRows))
 	assert.Nil(t, jLockedAgain)
 
 	// j3 should be still available as it is not gueron job
@@ -173,7 +169,7 @@ func Test_cleanupScheduledLeftovers(t *testing.T) {
 
 	jLockGueronJob, err := s.gueClient.LockJobByID(ctx, j4.ID)
 	require.Error(t, err)
-	assert.True(t, errors.Is(err, adapter.ErrNoRows))
+	assert.True(t, errors.Is(err, sql.ErrNoRows))
 	assert.Nil(t, jLockGueronJob)
 }
 
@@ -182,9 +178,9 @@ func Test_scheduleJobs(t *testing.T) {
 		t.Skip("skipping test in short mode")
 	}
 
-	ctx := context.Background()
-	pool := dbTest.OpenTestPoolPGXv5(t)
-	logger := zap.New(zaptest.NewLogger(t))
+	ctx := t.Context()
+	pool := openTestPoolPGXv5(t)
+	logger := slog.New(zapslog.NewHandler(zaptest.NewLogger(t).Core()))
 	queue := time.Now().Format(time.RFC3339Nano)
 	clk := clock.NewMock()
 
@@ -199,17 +195,17 @@ func Test_scheduleJobs(t *testing.T) {
 	now := time.Date(2022, 5, 8, 21, 27, 3, 0, time.UTC)
 	clk.Set(now)
 
-	tx, err := pool.Begin(ctx)
+	tx, err := pool.BeginTx(ctx, nil)
 	require.NoError(t, err)
 
 	schedulesHash := s.schedulesHash()
 	err = s.scheduleJobs(ctx, schedulesHash, tx)
 	require.NoError(t, err)
 
-	err = tx.Commit(ctx)
+	err = tx.Commit()
 	require.NoError(t, err)
 
-	rows, err := pool.Query(ctx, `SELECT job_type, run_at FROM gue_jobs WHERE queue = $1 ORDER BY run_at ASC`, queue)
+	rows, err := pool.QueryContext(ctx, `SELECT job_type, run_at FROM gue_jobs WHERE queue = $1 ORDER BY run_at ASC`, queue)
 	require.NoError(t, err)
 
 	jobs := make(map[string][]time.Time)
@@ -239,14 +235,14 @@ func Test_refreshSchedule(t *testing.T) {
 		t.Skip("skipping test in short mode")
 	}
 
-	ctx := context.Background()
-	pool := dbTest.OpenTestPoolPGXv5(t)
-	logger := zap.New(zaptest.NewLogger(t))
+	ctx := t.Context()
+	pool := openTestPoolPGXv5(t)
+	logger := slog.New(zapslog.NewHandler(zaptest.NewLogger(t).Core()))
 	queue := time.Now().Format(time.RFC3339Nano)
 	clk := clock.NewMock()
 	horizon := 10 * time.Minute
 
-	_, err := pool.Exec(ctx, `TRUNCATE gueron_meta`)
+	_, err := pool.ExecContext(ctx, `TRUNCATE gueron_meta`)
 	require.NoError(t, err)
 
 	s, err := NewScheduler(pool, WithLogger(logger), WithQueueName(queue), WithHorizon(horizon))
@@ -272,7 +268,7 @@ func Test_refreshSchedule(t *testing.T) {
 	)
 
 	err = pool.
-		QueryRow(ctx, `SELECT queue, hash, scheduled_at, horizon_at FROM gueron_meta`).
+		QueryRowContext(ctx, `SELECT queue, hash, scheduled_at, horizon_at FROM gueron_meta`).
 		Scan(&metaQueue, &metaHash, &metaScheduled, &metaHorizon)
 	require.NoError(t, err)
 
@@ -289,7 +285,7 @@ func Test_refreshSchedule(t *testing.T) {
 	require.NoError(t, err)
 
 	err = pool.
-		QueryRow(ctx, `SELECT queue, hash, scheduled_at, horizon_at FROM gueron_meta`).
+		QueryRowContext(ctx, `SELECT queue, hash, scheduled_at, horizon_at FROM gueron_meta`).
 		Scan(&metaQueue, &metaHash, &metaScheduled, &metaHorizon)
 	require.NoError(t, err)
 
@@ -307,7 +303,7 @@ func Test_refreshSchedule(t *testing.T) {
 	require.NoError(t, err)
 
 	err = pool.
-		QueryRow(ctx, `SELECT queue, hash, scheduled_at, horizon_at FROM gueron_meta`).
+		QueryRowContext(ctx, `SELECT queue, hash, scheduled_at, horizon_at FROM gueron_meta`).
 		Scan(&metaQueue, &metaHash, &metaScheduled, &metaHorizon)
 	require.NoError(t, err)
 
